Rename: 패키지명 수정 및 디버깅 코드 삭제
