+++ src/main/java/ojosama/talkak/category/config/CategoryInitializer.java
@@ -1,4 +1,4 @@
-package ojosama.talkak.category.util;
+package ojosama.talkak.category.config;
 
 import jakarta.annotation.PostConstruct;
 import java.util.Arrays;
